+++ src/main/java/com/alunos/alunosapi/AlunosApiApplication.java
@@ -5,9 +5,7 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
 public class AlunosApiApplication {
-
 	public static void main(String[] args) {
 		SpringApplication.run(AlunosApiApplication.class, args);
 	}
-
 }
