+++ src/main/resources/application.properties
@@ -1,5 +1,5 @@
 server.port = 8090
-spring.datasource.url=jdbc:mysql://127.0.0.1:3306/alunos?useTimezone=true&serverTimezone=UTC
+spring.datasource.url=jdbc:mysql://127.0.0.1:3306/alunos?allowPublicKeyRetrieval=true&sslMode=DISABLED&useSSL=false&useTimezone=true&serverTimezone=UTC
 spring.datasource.username=root
 spring.datasource.password=
 spring.jpa.hibernate.ddl-auto=update
